fixed some issues Favorites/HotelById/UserNotFoundException

+++ src/main/java/com/revature/DAOs/HotelDAO.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface HotelDAO extends JpaRepository<Hotel, Long> {
+public interface HotelDAO extends JpaRepository<Hotel, Integer> {
 
     // add custom methods WIP
